+++ src/main/java/nl/surfnet/bod/domain/validator/ReservationValidator.java
@@ -66,7 +66,7 @@ public class ReservationValidator implements Validator {
       return;
     }
 
-    Integer maxBandwidth = Math.min(sourcePort.getMaxBandwidth(), destinationPort.getMaxBandwidth());
+    long maxBandwidth = Math.min(sourcePort.getMaxBandwidth(), destinationPort.getMaxBandwidth());
     if (reservation.getBandwidth() > maxBandwidth) {
       errors.rejectValue("bandwidth", "validation.reservation.maxBandwidth", new Object[] { maxBandwidth },
           "bandwidth exceeded max");
